requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Data processing and visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration and logging
PyYAML>=6.0.0
tqdm>=4.62.0

# Development tools
black>=22.3.0
isort>=5.10.0
pylint>=2.14.0
pytest>=7.0.0
pytest-cov>=3.0.0

# Optional dependencies for notebooks
jupyter>=1.0.0
ipykernel>=6.0.0